Backend/requirements.txt
asgiref==3.7.2
Django==4.2.9
django-filter==23.5
django-grappelli==3.0.8
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.5.2
packaging==23.2
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
django-cors-headers==4.3.1



